server/weather_api_server/requirements.txt
grpcio
protobuf
requests
python-dotenv
redis